+++ package.json
@@ -16,9 +16,7 @@
     "@fortawesome/free-regular-svg-icons": "^6.5.2",
     "@fortawesome/free-solid-svg-icons": "^6.5.2",
     "@fortawesome/react-fontawesome": "^0.2.2",
-    "@mailchimp/mailchimp_marketing": "^3.0.80",
     "@next/third-parties": "^14.2.14",
-    "@types/mailchimp__mailchimp_marketing": "^3.0.20",
     "@vercel/analytics": "^1.3.1",
     "@vercel/speed-insights": "^1.0.12",
     "html-react-parser": "^5.1.12",
