+++ src/app/components/archive/index.tsx
@@ -6,7 +6,6 @@ import Link from "next/link"
 import { getPermalink, PageType } from "../../../../lib/utils"
 import Header, { HeaderType } from "../header"
 import Paper from "../paper"
-import DonatePopup from "../popup"
 
 export interface PromoProps {
   currentArticles: Articles[]
@@ -96,7 +95,6 @@ const ArchivePage = (props: ArchivePageProps) => {
           </div>
         </section>
       </div>
-      <DonatePopup />
     </Paper>
   )
 }
