+++ src/app/components/issueRail/coversPopup.tsx
@@ -1,39 +1,43 @@
 import parse from "html-react-parser"
 import { stripHtml } from "string-strip-html"
-import { usePopup } from "../popupProvider"
+import { usePopup } from "./popupProvider"
 import Image from "next/image"
 
 const CoversPopup = () => {
-  const { showPopup, images, popupType, setShowPopup } = usePopup() // Include popupType from context
+  const { showPopup, images } = usePopup()
 
-  if (!showPopup || popupType !== "coverImage") {
-    return null // Only show the popup if showPopup is true and popupType is "coverImage"
+  const { setShowPopup } = usePopup()
+
+  if (!showPopup) {
+    return null
   }
 
   const handleClick = async (e: React.MouseEvent<Element, MouseEvent>) => {
     e.preventDefault()
-    setShowPopup(false) // Hide the popup
+    setShowPopup(false)
   }
 
   const allCovers = images
     .filter((cover: any) => cover !== null)
     .map((cover: any, i: number) => {
       if (cover === null) {
-        return null
+        return <></>
       }
 
       const width = (cover.width * 500) / cover.height
       const height = (cover.height * width) / cover.width
       const src = `${process.env.NEXT_PUBLIC_IMAGE_PATH}${cover.filename_disk}`
       const alt = cover.caption ? `${stripHtml(cover.caption).result}` : "The Brooklyn Rail"
-
+      const caption = (
+        <>
+          <span className="uppercase font-medium">{`Cover ${i + 1}`}</span>: {parse(cover.caption)}
+        </>
+      )
       const key = `cover-${i}`
       return (
         <div key={`cover-${i}`} className="m-0 py-2 w-card-lg tablet-lg:w-card-lg flex-none">
           <Image src={src} width={width} height={height} alt={alt} sizes="35vw" />
-          <figcaption className="pt-2 text-white text-xs">
-            <span className="uppercase font-medium">{`Cover ${i + 1}`}</span>: {parse(cover.caption)}
-          </figcaption>
+          <figcaption className="pt-2 text-white text-xs">{caption}</figcaption>
         </div>
       )
     })
@@ -47,5 +51,4 @@ const CoversPopup = () => {
     </div>
   )
 }
-
 export default CoversPopup
